+++ lib/titles.js
@@ -1,4 +1,4 @@
-import productConfig from "../config/print-mapping.js";
+import productConfig from "../config/product-mapping.js";
 
 const postalDeliveryAllowedTitles = [ "dn", "paf", "plg" ];
